feat: Adding Gson and Moshi artifact
