+++ emojify/build.gradle.kts
@@ -1,3 +1,5 @@
+import io.wax911.emoji.buildSrc.Libraries
+
 plugins {
     id("io.wax911.emojify")
 }
@@ -5,7 +7,7 @@ plugins {
 dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.startup.runtime)
-    implementation(project(":contract"))
+    implementation(project(Libraries.AniTrend.Emojify.contract))
 }
 
 android {
